+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_en.dart
@@ -659,4 +659,7 @@ class StreamChatLocalizationsEn extends GlobalStreamChatLocalizations {
 
   @override
   String pollSomeoneCreatedText(String username) => '$username created';
+
+  @override
+  String get draftLabel => 'Draft';
 }