+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -669,4 +669,7 @@ Limite de pièces jointes dépassée : il n'est pas possible d'ajouter plus de $
 
   @override
   String pollSomeoneCreatedText(String username) => '$username a créé';
+
+  @override
+  String get draftLabel => 'Brouillon';
 }