+++ packages/stream_chat_flutter_core/CHANGELOG.md
@@ -1,3 +1,9 @@
+## Upcoming
+
+✅ Added
+
+- Added `StreamDraftListController` to manage the list of draft messages.
+
 ## 9.8.0
 
 ✅ Added